Merge dev into master

+++ src/main/java/net/maritimeconnectivity/serviceregistry/services/XmlService.java
@@ -17,13 +17,23 @@
 package net.maritimeconnectivity.serviceregistry.services;
 
 import lombok.extern.slf4j.Slf4j;
+import net.maritimeconnectivity.serviceregistry.exceptions.DataNotFoundException;
 import net.maritimeconnectivity.serviceregistry.models.domain.Xml;
+import net.maritimeconnectivity.serviceregistry.models.domain.enums.G1128Schemas;
 import net.maritimeconnectivity.serviceregistry.repos.XmlRepo;
+import net.maritimeconnectivity.serviceregistry.utils.G1128Utils;
+import net.maritimeconnectivity.serviceregistry.utils.XmlUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.xml.sax.SAXException;
+
+import javax.xml.bind.JAXBException;
+import java.io.IOException;
+import java.util.Optional;
 
 /**
  * Service Implementation for managing Xml.
@@ -35,20 +45,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class XmlService {
 
+    /**
+     * The XML Repository.
+     */
     @Autowired
     private XmlRepo xmlRepo;
-
-    /**
-     * Save a xml.
-     *
-     * @param xml the entity to save
-     * @return the persisted entity
-     */
-    public Xml save(Xml xml) {
-        log.debug("Request to save Xml : {}", xml);
-        Xml result = this.xmlRepo.save(xml);
-        return result;
-    }
 
     /**
      *  Get all the xmls.
@@ -59,31 +60,67 @@ public class XmlService {
     @Transactional(readOnly = true)
     public Page<Xml> findAll(Pageable pageable) {
         log.debug("Request to get all Xmls");
-        Page<Xml> result = this.xmlRepo.findAll(pageable);
-        return result;
+        return this.xmlRepo.findAll(pageable);
     }
 
     /**
-     *  Get one xml by id.
+     *  Get one xml by ID.
      *
-     *  @param id the id of the entity
+     *  @param id       the ID of the entity
      *  @return the entity
      */
     @Transactional(readOnly = true)
-    public Xml findOne(Long id) {
+    public Xml findOne(Long id) throws DataNotFoundException {
         log.debug("Request to get Xml : {}", id);
-        Xml xml = this.xmlRepo.findById(id).orElse(null);
-        return xml;
+        return this.xmlRepo.findById(id)
+                .orElseThrow(() -> new DataNotFoundException("", null));
     }
 
     /**
-     *  Delete the  xml by id.
+     * Save a xml.
      *
-     *  @param id the id of the entity
+     * @param xml       the entity to save
+     * @return the persisted entity
      */
-    public void delete(Long id) {
+    @Transactional
+    public Xml save(Xml xml) {
+        log.debug("Request to save Xml : {}", xml);
+        return this.xmlRepo.save(xml);
+    }
+
+    /**
+     *  Delete the  xml by ID.
+     *
+     *  @param id       the ID of the entity
+     */
+    @Transactional(propagation = Propagation.NESTED)
+    public void delete(Long id) throws DataNotFoundException {
         log.debug("Request to delete Xml : {}", id);
-        this.xmlRepo.deleteById(id);
+        if(this.xmlRepo.existsById(id)) {
+            this.xmlRepo.deleteById(id);
+        } else {
+            throw new DataNotFoundException("No xml found for the provided ID", null);
+        }
+    }
+
+    /**
+     * This function accepts an XML content as an input, as well as the G1128
+     * specification schema, in order to validate whether the input is actually
+     * correct.
+     *
+     * @param content   the XML content
+     * @param schema    the G1128 schema to validate the content with
+     * @return the generated G1128 specification object
+     */
+    public Object validate(String content, G1128Schemas schema) throws IOException, SAXException, JAXBException {
+        // Make sure we have a valid G1128 schema class to work with
+        Class schemaClass = Optional.of(schema)
+                .map(G1128Schemas::getSchemaClass)
+                .orElseThrow(() -> new DataNotFoundException("Invalid G1128 schema selection", null));
+        // Make sure it's a valid XML
+        XmlUtil.validateXml(content, G1128Utils.SOURCES_LIST);
+        // And parse the G1128 content
+        return new G1128Utils<>(schemaClass).unmarshallG1128(content);
     }
 
 }
